Change one test to fit new paradigm
This test was a simple case in the old paradigm but covers several cases in the new paradigm, so we've split it out into three tests. The original test fails because exclusion is not yet implmeneted. One of the new tests fails because not yet implemented. One test passes, yay!

(The point is that in the old paradigm, exhausted votes were not explicitly dealt with - they need to be in new paradigm)