Move Dom from continuing to provisionally elected
Initially we thought that if the last remaining candidate doesn't reach the quota then it was a failed election. Now, based on the paper we are using as definitive, if a candidate is eliminated such that the remaining number of candidates = the number of vacancies, then all those candidates are elected regardless of whether they meet the quota. This means that this test doesn't need the extra round as Dom is elected at the end of the second round.