Implement bulk exclusion
We either delete a bunch of low ones if that's appropriate (will only happen once) or the lowest. Still raise an error if tied, but after bulk exclusion, so not on, e.g. tied 0s.

A bunch of tests fail now but they fail because, e.g. bulk exclusion kills so many that someone now goes into provisionally elected.